moved reusable to shared
