+++ src/components/ExternalNotes.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { setNotes } from "../apiHooks/setNotes";
 import { toast } from "react-toastify";
-import { Text, ToastErrorBox } from "./Reusble";
+import { Text, ToastErrorBox } from "../../shared/Reusble.jsx";
 import { ToastError } from "../Utils/functions";
 
 export default function ExternalNotes({
