+++ src/pages/ContactPage/ContactUs.jsx
@@ -5,7 +5,7 @@ import * as Yup from "yup";
 import { errorMsgClass, fieldClass } from "../../Utils/HardCoded";
 import sendContactUs from "../../apiHooks/sendContactUs";
 import { toast } from "react-toastify";
-import { ToastBox } from "../../components/Reusble";
+import { ToastBox } from "../../../shared/Reusble.jsx";
 
 export default function ContactUs() {
   const [errorMsg, setErrorMsg] = useState(false);
