+++ src/pages/ContactPage/SuggestingNewQueries.jsx
@@ -5,7 +5,7 @@ import * as Yup from "yup";
 import { errorMsgClass, fieldClass } from "../../Utils/HardCoded";
 import sendSuggestNewQuery from "../../apiHooks/sendSuggestNewQuery";
 import { toast } from "react-toastify";
-import { ToastBox } from "../../components/Reusble";
+import { ToastBox } from "../../../shared/Reusble.jsx";
 
 export default function SuggestingNewQueries() {
   const [errorMsg, setErrorMsg] = useState(false);
