+++ src/pages/ParametersDistributionBar/ParametersDistributionBar.jsx
@@ -10,7 +10,7 @@ import {
   TopGraphText,
   CSV,
   Reset,
-} from "../../components/Reusble";
+} from "../../../shared/Reusble.jsx";
 import getExperimentsGraphs from "../../apiHooks/getExperimentsGraphs";
 import Plot from "react-plotly.js";
 import {
