+++ src/pages/TheoryDriven/TheoryDriven.jsx
@@ -18,7 +18,7 @@ import {
   Text,
   TopGraphText,
   TypeOfConsciousnessFilter,
-} from "../../components/Reusble";
+} from "../../../shared/Reusble.jsx";
 import getExperimentsGraphs from "../../apiHooks/getExperimentsGraphs";
 import Plot from "react-plotly.js";
 import Toggle from "../../components/Toggle";
