+++ src/pages/Timing/Timings.jsx
@@ -11,7 +11,7 @@ import {
   TheoryDrivenFilter,
   TopGraphText,
   TypeOfConsciousnessFilter,
-} from "../../components/Reusble";
+} from "../../../shared/Reusble.jsx";
 import Plot from "react-plotly.js";
 import {
   plotConfig,
