+++ src/pages/UploadNewPaper/ExperimentsSection/BasicClassification.jsx
@@ -6,7 +6,7 @@ import {
   Button,
   ToastBox,
   ToastErrorBox,
-} from "../../../components/Reusble";
+} from "../../../../shared/Reusble.jsx";
 import * as Yup from "yup";
 import Select from "react-select";
 import {
