+++ src/pages/UploadNewPaper/ExperimentsSection/ConsciousnessMeasures.jsx
@@ -6,7 +6,7 @@ import {
   TrashButton,
   CustomSelect,
   CircledIndex,
-} from "../../../components/Reusble";
+} from "../../../../shared/Reusble.jsx";
 import { useEffect, useState } from "react";
 import {
   DeleteClassificationField,
