+++ src/pages/UploadNewPaper/ExperimentsSection/ExperimentForm.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Text } from "../../../components/Reusble";
+import { Text } from "../../../../shared/Reusble.jsx";
 import { useQuery } from "@tanstack/react-query";
 import getExtraConfig from "../../../apiHooks/getExtraConfig";
