+++ src/pages/UploadNewPaper/NewPaperForm.jsx
@@ -9,7 +9,7 @@ import {
   Spacer,
   ToastBox,
   ToastErrorBox,
-} from "../../components/Reusble";
+} from "../../../shared/Reusble.jsx";
 import { errorMsgClass } from "../../Utils/HardCoded";
 import * as Yup from "yup";
 import MultiSelect from "../../components/SelectField";
