+++ src/pages/UploadNewPaper/PaperSection.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import ProgressComponent from "./ProgressComponent";
-import { Spacer, Text } from "../../components/Reusble";
+import { Spacer, Text } from "../../../shared/Reusble.jsx";
 import ExperimentsBox from "./ExperimentsBox";
 import { countries } from "countries-list";
